Add license header to new modules
