+++ src/main.rs
@@ -9,6 +9,10 @@ use clap::Parser;
 use std::fs::File;
 use std::io::{self, BufRead, BufReader, BufWriter, Write};
 use std::path::Path;
+mod header;
+mod scrub;
+use crate::header::{detect_header, ask_yes_no_question};
+use crate::scrub::{ChangeInfo, StreamState, VerboseCommentType, process_line_streaming};
 
 #[derive(Parser, Debug)]
 #[clap(name = "rustscrub", author = "Volker Schwaberow <volker@schwaberow.de>", version, about = "RustScrub: Removes comments from Rust files.", long_about = None)]
@@ -28,313 +32,6 @@ struct Args {
     #[clap(short, long, action = clap::ArgAction::SetTrue)]
     dry_run: bool,
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum VerboseCommentType {
-    Line,
-    Block,
-}
-
-#[derive(Debug, Clone)]
-struct ChangeInfo {
-    start_line: usize, 
-    end_line: usize,   
-    comment_type: VerboseCommentType,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum State {
-    Normal,
-    LineComment,
-    BlockComment,
-    StringLiteral,
-    StringEscape,
-    CharLiteral,
-    CharEscape,
-    InRawString,
-}
-
-#[derive(Debug, Clone, Copy)]
-struct StreamState {
-    current_parse_state: State,
-    raw_string_hash_count: usize,
-    active_block_comment_start_line: Option<usize>, 
-    is_processing_full_line_comment: bool, 
-}
-
-impl Default for StreamState {
-    fn default() -> Self {
-        StreamState {
-            current_parse_state: State::Normal,
-            raw_string_hash_count: 0,
-            active_block_comment_start_line: None,
-            is_processing_full_line_comment: false, 
-        }
-    }
-}
-
-fn detect_header(file_path: &Path) -> Result<(usize, String), String> {
-    let file = File::open(file_path)
-        .map_err(|e| format!("Failed to open file for header detection: {}", e))?;
-    
-    let reader = BufReader::new(file);
-    let mut lines = Vec::new();
-    let mut line_count = 0;
-    let mut in_header = true;
-    
-    let mut saw_code = false;
-    let mut saw_normal_comment = false;
-    let mut empty_line_count = 0;
-    
-    const MAX_PREVIEW_LINES: usize = 10;
-    const MAX_HEADER_SIZE: usize = 50;  
-    
-    for line in reader.lines() {
-        let line = line.map_err(|e| format!("Failed to read line during header detection: {}", e))?;
-        line_count += 1;
-        
-        if line_count <= MAX_PREVIEW_LINES {
-            lines.push(line.clone());
-        }
-        
-        let trimmed = line.trim();
-        
-        if trimmed.is_empty() {
-            empty_line_count += 1;
-            if empty_line_count > 2 && saw_normal_comment {
-                in_header = false;
-                break;
-            }
-            continue;
-        } else {
-            empty_line_count = 0;
-        }
-        
-        if trimmed.starts_with("#![") {
-            saw_normal_comment = true;
-            continue;
-        }
-        
-        if trimmed.starts_with("//!") || trimmed.starts_with("///") {
-            saw_normal_comment = true;
-            continue;
-        }
-        
-        if trimmed.starts_with("//") || trimmed.starts_with("/*") {
-            saw_normal_comment = true;
-            continue;
-        }
-        
-        if trimmed.starts_with("use ") || trimmed.starts_with("mod ") || 
-           trimmed.starts_with("pub ") || trimmed.starts_with("fn ") ||
-           trimmed.starts_with("struct ") || trimmed.starts_with("enum ") ||
-           trimmed.starts_with("impl ") || trimmed.starts_with("trait ") {
-            saw_code = true;
-            in_header = false;
-            break;
-        }
-        
-        if line_count > 3 && saw_normal_comment {
-            in_header = false;
-            break;
-        }
-        
-        if line_count > MAX_HEADER_SIZE {
-            in_header = false;
-            break;
-        }
-    }
-    
-    let header_lines = if saw_code && !in_header {
-        line_count - 1  
-    } else if saw_normal_comment {
-        line_count  
-    } else {
-        0  
-    };
-    
-    let preview = if !lines.is_empty() {
-        let preview_text = lines.join("\n");
-        if lines.len() < line_count {
-            format!("{}\n... ({} more lines)", preview_text, line_count - lines.len())
-        } else {
-            preview_text
-        }
-    } else {
-        "".to_string()
-    };
-    
-    Ok((header_lines, preview))
-}
-
-fn ask_yes_no_question(question: &str) -> bool {
-    use std::io::{stdin, stdout};
-    
-    print!("{} [y/N]: ", question);
-    stdout().flush().unwrap_or(());
-    
-    let mut response = String::new();
-    
-    if stdin().read_line(&mut response).is_err() {
-        return false;
-    }
-    
-    let response = response.trim().to_lowercase();
-    response == "y" || response == "yes"
-}
-
-fn process_line_streaming(
-    line_content: &str,
-    original_line_num: usize,
-    stream_state: &mut StreamState,
-) -> (String, Vec<ChangeInfo>) {
-    let mut output_segment = String::with_capacity(line_content.len());
-    let mut chars = line_content.chars().peekable();
-    let mut line_changes = Vec::new();
-
-
-    while let Some(current_char) = chars.next() {
-        match stream_state.current_parse_state {
-            State::Normal => {
-                match current_char {
-                    '/' => {
-                        if chars.peek() == Some(&'/') {
-                            chars.next(); 
-                            if output_segment.trim().is_empty() {
-                                output_segment.clear(); 
-                                stream_state.is_processing_full_line_comment = true;
-                            } else {
-                                stream_state.is_processing_full_line_comment = false;
-                            }
-                            stream_state.current_parse_state = State::LineComment;
-                            line_changes.push(ChangeInfo {
-                                start_line: original_line_num,
-                                end_line: original_line_num,
-                                comment_type: VerboseCommentType::Line,
-                            });
-                        } else if chars.peek() == Some(&'*') {
-                            chars.next(); 
-                            stream_state.current_parse_state = State::BlockComment;
-                            if stream_state.active_block_comment_start_line.is_none() {
-                                stream_state.active_block_comment_start_line = Some(original_line_num);
-                            }
-                        } else {
-                            output_segment.push(current_char);
-                        }
-                    }
-                    '"' => {
-                        output_segment.push(current_char);
-                        stream_state.current_parse_state = State::StringLiteral;
-                    }
-                    '\'' => {
-                        output_segment.push(current_char);
-                        stream_state.current_parse_state = State::CharLiteral;
-                    }
-                    'r' => {
-                        let mut temp_hashes = 0;
-                        let mut prefix_buffer = String::from('r');
-                        while let Some(&'#') = chars.peek() {
-                            prefix_buffer.push(chars.next().unwrap());
-                            temp_hashes += 1;
-                        }
-                        if let Some(&'"') = chars.peek() {
-                            stream_state.raw_string_hash_count = temp_hashes;
-                            output_segment.push_str(&prefix_buffer);
-                            output_segment.push(chars.next().unwrap()); 
-                            stream_state.current_parse_state = State::InRawString;
-                        } else {
-                            output_segment.push_str(&prefix_buffer);
-                        }
-                    }
-                    _ => { 
-                        output_segment.push(current_char);
-                    }
-                }
-            }
-            State::LineComment => {
-                if current_char == '\n' { 
-                    if !stream_state.is_processing_full_line_comment {
-                        output_segment.push(current_char); 
-                    }
-                    stream_state.current_parse_state = State::Normal; 
-                    stream_state.is_processing_full_line_comment = false; 
-                }
-            }
-            State::BlockComment => {
-                if current_char == '*' && chars.peek() == Some(&'/') {
-                    chars.next(); 
-                    stream_state.current_parse_state = State::Normal; 
-                    if let Some(start_line) = stream_state.active_block_comment_start_line {
-                        line_changes.push(ChangeInfo {
-                            start_line,
-                            end_line: original_line_num,
-                            comment_type: VerboseCommentType::Block,
-                        });
-                        stream_state.active_block_comment_start_line = None; 
-                    }
-                }
-            }
-            State::StringLiteral => {
-                output_segment.push(current_char);
-                match current_char {
-                    '\\' => stream_state.current_parse_state = State::StringEscape,
-                    '"' => stream_state.current_parse_state = State::Normal,
-                    _ => {} 
-                }
-            }
-            State::StringEscape => {
-                output_segment.push(current_char);
-                stream_state.current_parse_state = State::StringLiteral;
-            }
-            State::CharLiteral => {
-                output_segment.push(current_char);
-                match current_char {
-                    '\\' => stream_state.current_parse_state = State::CharEscape,
-                    '\'' => stream_state.current_parse_state = State::Normal,
-                    _ => {}
-                }
-            }
-            State::CharEscape => {
-                output_segment.push(current_char);
-                stream_state.current_parse_state = State::CharLiteral;
-            }
-            State::InRawString => {
-                output_segment.push(current_char);
-                if current_char == '"' {
-                    let mut closing_hashes_candidate = String::new();
-                    let mut hashes_found = 0;
-                    let mut is_proper_closing_sequence = true;
-
-                    if stream_state.raw_string_hash_count > 0 {
-                        for _ in 0..stream_state.raw_string_hash_count {
-                            if let Some(&peeked_char) = chars.peek() {
-                                if peeked_char == '#' {
-                                    closing_hashes_candidate.push(chars.next().unwrap());
-                                    hashes_found += 1;
-                                } else {
-                                    is_proper_closing_sequence = false;
-                                    break;
-                                }
-                            } else {
-                                is_proper_closing_sequence = false;
-                                break;
-                            }
-                        }
-                    }
-                    if is_proper_closing_sequence && hashes_found == stream_state.raw_string_hash_count {
-                        output_segment.push_str(&closing_hashes_candidate);
-                        stream_state.current_parse_state = State::Normal;
-                        stream_state.raw_string_hash_count = 0;
-                    } else {
-                        output_segment.push_str(&closing_hashes_candidate);
-                    }
-                }
-            }
-        }
-    }
-    (output_segment, line_changes)
-}
-
 
 fn main() -> Result<(), String> {
     let mut args = Args::parse();
